refactor(firehose-protos): reorganize ethereum v1 module
